+++ nautilus_trader/model/events/order.pyx
@@ -13,7 +13,6 @@
 #  limitations under the License.
 # -------------------------------------------------------------------------------------------------
 
-import json
 from typing import Any
 
 from libc.stdint cimport uint64_t
@@ -527,7 +526,7 @@ cdef class OrderInitialized(OrderEvent):
         cdef str parent_order_id_str = values["parent_order_id"]
         cdef str exec_algorithm_id_str = values["exec_algorithm_id"]
         cdef str exec_spawn_id_str = values["exec_spawn_id"]
-        exec_algorithm_params_json = values["exec_algorithm_params"]
+        cdef dict exec_algorithm_params = values["exec_algorithm_params"]
         return OrderInitialized(
             trader_id=TraderId(values["trader_id"]),
             strategy_id=StrategyId(values["strategy_id"]),
@@ -540,7 +539,7 @@ cdef class OrderInitialized(OrderEvent):
             post_only=values["post_only"],
             reduce_only=values["reduce_only"],
             quote_quantity=values["quote_quantity"],
-            options=json.loads(values["options"]),  # Using vanilla json due mixed schema types
+            options=values["options"],
             emulation_trigger=trigger_type_from_str(values["emulation_trigger"]),
             trigger_instrument_id=InstrumentId.from_str_c(trigger_instrument_id) if trigger_instrument_id is not None else None,
             contingency_type=contingency_type_from_str(values["contingency_type"]),
@@ -548,7 +547,7 @@ cdef class OrderInitialized(OrderEvent):
             linked_order_ids=[ClientOrderId(o_str) for o_str in linked_order_ids_str.split(",")] if linked_order_ids_str is not None else None,
             parent_order_id=ClientOrderId(parent_order_id_str) if parent_order_id_str is not None else None,
             exec_algorithm_id=ExecAlgorithmId(exec_algorithm_id_str) if exec_algorithm_id_str is not None else None,
-            exec_algorithm_params=json.loads(exec_algorithm_params_json) if exec_algorithm_params_json is not None else None,
+            exec_algorithm_params=exec_algorithm_params,
             exec_spawn_id=ClientOrderId(exec_spawn_id_str) if exec_spawn_id_str is not None else None,
             tags=values["tags"],
             event_id=UUID4(values["event_id"]),
@@ -573,7 +572,7 @@ cdef class OrderInitialized(OrderEvent):
             "post_only": obj.post_only,
             "reduce_only": obj.reduce_only,
             "quote_quantity": obj.quote_quantity,
-            "options": json.dumps(obj.options),  # Using vanilla json due mixed schema types
+            "options": obj.options,
             "emulation_trigger": trigger_type_to_str(obj.emulation_trigger),
             "trigger_instrument_id": obj.trigger_instrument_id.value if obj.trigger_instrument_id is not None else None,
             "contingency_type": contingency_type_to_str(obj.contingency_type),
@@ -581,7 +580,7 @@ cdef class OrderInitialized(OrderEvent):
             "linked_order_ids": ",".join([o.value for o in obj.linked_order_ids]) if obj.linked_order_ids is not None else None,  # noqa
             "parent_order_id": obj.parent_order_id.value if obj.parent_order_id is not None else None,
             "exec_algorithm_id": obj.exec_algorithm_id.value if obj.exec_algorithm_id is not None else None,
-            "exec_algorithm_params": json.dumps(obj.exec_algorithm_params) if obj.exec_algorithm_params else None,
+            "exec_algorithm_params": obj.exec_algorithm_params,
             "exec_spawn_id": obj.exec_spawn_id.value if obj.exec_spawn_id is not None else None,
             "tags": obj.tags,
             "event_id": obj.id.value,